Change to mol_com, add density_from_frame
mol_com functions now returns the mass of each molecule and their
implementation for calculating the center of mass was improved. The
function density_from_frame was created, along with a unit test that
checks for each of its centering methods. Unit testing naming
conventions also changed, so now test_<variable> is the result to check
in the test. Changed docstrings for functions as well.